refactor(tui): split setup modules
